`#wip` range join tests
`testOnly com.gitlab.leafty.test.misc.RangeJoinSpec` works

+++ src/test/scala/com/gitlab/leafty/test/misc/package.scala
@@ -2,8 +2,8 @@ package com.gitlab.leafty.test
 
 package object misc {
 
-  case class Trn(amount: Integer, time: Integer)
+  case class Trn(amount: Integer, time: Long)
 
-  case class Range(start: Integer, end: Integer)
+  case class Range(start: Long, end: Long)
 
 }
